+++ tests/core/operator/test_operator2.py
@@ -25,7 +25,7 @@ from scipy.sparse import csr_matrix
 
 import pennylane as qp
 from pennylane.core.operator import Operator2, StatePrepBase2
-from pennylane.core.queuing import AnnotatedQueue
+from pennylane.core.queuing import AnnotatedQueue, apply
 from pennylane.exceptions import (
     AdjointUndefinedError,
     DecompositionUndefinedError,
@@ -2440,3 +2440,24 @@ class TestLegacyCompatibilityViews:
         assert "phi" not in op.hyperparameters
         assert "wires" not in op.hyperparameters
         assert "aux_wires" not in op.hyperparameters
+
+
+class TestApply:
+
+    @pytest.mark.parametrize("op2", [DynOp(1.0, wires=0), FullOp(0.3, "lbl", [1.0, 2.0], wires=0)])
+    def test_queues_without_capture(self, op2):
+        """Tests that Operator2 can queue like Operator1 with capture disabled."""
+
+        with AnnotatedQueue() as q:
+            apply(op2)
+
+        assert len(q.queue) == 1
+        assert q.queue[0] == op2
+
+    def test_raises_capture_disabled_outside_queueing_context(self):
+        """Tests that outside a queuing context and without capture enabled, apply() raises when given an Operator2."""
+
+        with pytest.raises(
+            RuntimeError, match="No queuing context available to append operation to"
+        ):
+            apply(DynOp(1.0, wires=0))
